Generate video output at the beginning of VSYNC rather than at the end of it. This makes video decoding use  correct values from the Gate Array and the CRTC.
This "fixes" Target Renegade. Now it runs fine again, except for that the whole frame is decoded using video mode 1 (play area should use mode 0). This is exactly the expected result as we are decoding the whole frame at the end. Will be fixed by decoding one scan line at a time.